EnemySpawner script is now dynamic
The range of location that they spawn from and their rotation are now dynamic for each instance of the script.

For a practice to keep it organized I would advise attaching a new instance of the script to each "portal" that we make/make appear. You can attach scripts to new objects programmatically during runtime if that's what we want to do.